+++ app/src/main/java/com/techcourse/DispatcherServlet.java
@@ -7,8 +7,8 @@ import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.mvc.tobe.AnnotationHandlerMapping;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerMapping;
+import webmvc.org.springframework.web.servlet.mvc.tobe.mapping.AnnotationHandlerMapping;
+import webmvc.org.springframework.web.servlet.mvc.tobe.mapping.HandlerMapping;
 import webmvc.org.springframework.web.servlet.mvc.tobe.adapter.ControllerHandlerAdapter;
 import webmvc.org.springframework.web.servlet.mvc.tobe.adapter.HandlerAdapter;
 import webmvc.org.springframework.web.servlet.mvc.tobe.adapter.HandlerExecutionHandlerAdapter;
